Add a Death class which is a Scene
The game is to consist of a Scene called Death, therefore a Death
class was created that inherits from Scene. The class consists of
an array(list) that contains different 'quips'. When a player enters
the Death Scene a quip is randomly selected from the list and
displayed to the user.