Add Central Corridor Scene class
The game begins in a Scene that is a Central Corrdidor, therefore
this class needed to be created. In this scene there are three
different actions you can take after the inital prompt upon entering
the scene. The actions are shoot, dodge, or tell a joke. And if-else
statement was created to handle all three of the actions including
and else for when a correct action is not chosen, which returns you
to the Central Corridor. The action of tell a joke sends you to the
Laser Weapon Armory Scene (TBC).