Add Laser Weapon Armory Scene
The Central Corridor Scene called for a Laser Weapon Armory Scene
when 'tell a joke' is the action entered. There is an initial entry
called when the LWA Scene is entered. Inside this scene the player
must enter a 3 digit code to unlock another door (Scene). Therefore
a class was created for this Scene. The player gets 10 guesses for
the correct code, but if never guessed the user is sent to the Death
Scene. If the user does correctly guess the code they are sent to
another Scene called The Bridge.