Add Bridge Scene
The player advances to The Bridge Scene when they guess the code
correctly inside of the LWA Scene, therefore this scene needed to
be created. Inside this scene there are only two actions to take,
either 'throw the bomb' or 'slowly place the bomb'. When 'throw
the bomb' is chosen the player is sent to the Death Scene. When the
player chooses 'slowly place the bomb' the user is sent to the
Escape Pod Scene. Otherwise The Bridge scene is re-rendered.