Add Escape Pod Scene
The last scene (The Bridge) calls for an Escape Pod scene when the
player chooses to 'slowly place the bomb', therefore an Escape Pod
Scene was created. Here the player is prompted to choose from 5
escape pods. If the user chooses what is randomly designated as the
'good pod' (the else part of the if-statement) then the player has
won and designated as 'finished' which calls the Finished class.
If the user does not choose the 'good pod' then they are sent to
the Death Scene.