Add a Map to handle all of the Scene callings
During each of the Scenes a word is returned that describes which
Scene to go to next, but there was nothing directing the program
to actually call these classes, so a Map class was created. The
Map contains a scenes dictionary which uses the returned words from
the scenes as the keys and calling their respective class as the
value. The initialization of the Map class the start scene is set.
The next scene method takes the value obtained from the scene class
called and returns it. The opening scene method returns the next
scene in the game.