Add game play
All of the necessary functions for the game were created, but the
game wasn't being called. So the actions were added to create a map
and then the map was used in an instance of an Engine to create a
game. The play method was called on the game to begin the actual
gameplay.