Define recursive method for Fibonacci
The first fibonacci method was continuously running through the
method, but wasn't very efficient for if a large number was entered.
Therefore a recursive method was written where the method is
continuously called inside of itself until the answer is found and
then exits out once its done and returns that number.